requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
bidict==0.23.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
cryptography==44.0.2
dashscope==1.23.4
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
json5==0.12.0
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mcp==1.9.4
multidict==6.4.4
nest-asyncio==1.6.0
openai==1.86.0
oracledb==3.0.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pip==25.0.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pyzmq==26.3.0
qwen-agent==0.0.26
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.25.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.47.0
tiktoken==0.9.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
uv==0.7.5
uvicorn==0.34.3
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
